Add Ignore field
The addition of the Field type allows files with meaningless data (e.g. Excel
formulas) to be accounted for, and summarily dismissed. This is particularly
useful in the case of empty row checks, as it allows us to consider the row
empty if there are no values in any non-Ignore field.