Correcting choice rule prepare method
When feeding cells with comparison dictionaries (as occurs with row level
pass of cell values), the prepare method was failing, causing Choice rule
evaluation to always fail. This was not captured in original testing, as
there were only tests which directly passed to the Cell, rather than the
row based method which includes comparisons. Additional tests have been
added to address this case.