Latest (#35)
* Checkpoint exception updates

* Checkpoint on rule generated exceptions

* Change base method name to better reflect parent object

Class exception was ambiguous. Rule exception gets right to the point.

* Update minor version to reflect public API changes

* File name patterns not currently implemented

Removing, as it was untested code.

* Remove empty rule check as it is not fully implemented or tested

* Remove list if condition

The error is never a list. This doesn't happen in the code, and this method is semi-private, so it should never happen.

* Correct typehint for modified error flatten method

* Remove unnecessary recursive class checking

* Add proper support for column rule error reporting

Everything was previously reported as a cell error.

* Add very silly test for exception message structure generation

* Remove unused method

* Remove unused method

* Use more specific Exception

* Extend test coverage to 100%

* Add codecov configuration to ignore setup.py for testing coverage

This always gets tested as part of the install step.