v0.4.0 (#41)
* Checkpoint exception updates

* Checkpoint on rule generated exceptions

* Change base method name to better reflect parent object

Class exception was ambiguous. Rule exception gets right to the point.

* Update minor version to reflect public API changes

* File name patterns not currently implemented

Removing, as it was untested code.

* Remove empty rule check as it is not fully implemented or tested

* Remove list if condition

The error is never a list. This doesn't happen in the code, and this method is semi-private, so it should never happen.

* Correct typehint for modified error flatten method

* Remove unnecessary recursive class checking

* Add proper support for column rule error reporting

Everything was previously reported as a cell error.

* Add very silly test for exception message structure generation

* Extend test coverage to 100%

* Cli menu (#37)

* Menu checkpointing

* Checkpoint menu

* Mock! YEAH! Ing! YEAH!

* Got conditional import to work for mock Markdown

I think...

* Stablize menu tests

* Complete test coverage of documentation formats

* Resolve built-in name warning

* Add pytest mock to Testing Extra

* Updates

* Complete code coverage of menu

* Suppress code inspection warnings

Was warning about duplicated code, but abstracting the code in a unit test
would not make it more clear; it would likely make the test conditions less
clear.

* Update object import structure

We want the menu function to available at the root level of the package.

* Updating docstrings and sphinx docs

* File classes (#39)

Add dedicated file classes to handle checking of different file types.

* Latest (#38)

* Checkpoint exception updates

* Checkpoint on rule generated exceptions

* Change base method name to better reflect parent object

Class exception was ambiguous. Rule exception gets right to the point.

* Update minor version to reflect public API changes

* File name patterns not currently implemented

Removing, as it was untested code.

* Remove empty rule check as it is not fully implemented or tested

* Remove list if condition

The error is never a list. This doesn't happen in the code, and this method is semi-private, so it should never happen.

* Correct typehint for modified error flatten method

* Remove unnecessary recursive class checking

* Add proper support for column rule error reporting

Everything was previously reported as a cell error.

* Add very silly test for exception message structure generation

* Remove unused method

* Remove unused method

* Use more specific Exception

* Extend test coverage to 100%

* Add codecov configuration to ignore setup.py for testing coverage

This always gets tested as part of the install step.

* Cli menu (#37)

* Menu checkpointing

* Checkpoint menu

* Mock! YEAH! Ing! YEAH!

* Got conditional import to work for mock Markdown

I think...

* Stablize menu tests

* Complete test coverage of documentation formats

* Resolve built-in name warning

* Add pytest mock to Testing Extra

* Updates

* Complete code coverage of menu

* Suppress code inspection warnings

Was warning about duplicated code, but abstracting the code in a unit test
would not make it more clear; it would likely make the test conditions less
clear.

* Update object import structure

We want the menu function to available at the root level of the package.

* Add menu object to example

This is the preferred way of sharing the validation script.

* Updating docstrings and sphinx docs

* Set up file validators as dedicated classes

* Modify unit tests to accommodate dedicated file classes

* Add docs for file type classes

* Complete coverage

* Add coverage for openpyxl cases

* Correcting docstrings

* Add self reference to sphinx index

* Remove unreliable test of Excel class import

* Abstract module mocker

* Correct docstring

* Debug mode (#40)

* Correct docstring

* Add debug mode with instructions in exception module method

* Field strip (#42)

Add strip kwarg to Field constructor

This will trigger pre-processing on any values to be checked by that field, which will remove leading and trailing whitespaces.

* Add field parameter to perform strip on values

Still need to complete test coverage

* Add check of debug which skips pre-processing

* Add docstring of changes

* Codecov thought we we're executing the other method

I guess it's not perfect

* Header stripper (#43)

* checkpoint header stripper

* Remove unused extra param

* Corrected tests and header mode functions