Rework row rules to pass unit tests
I made maybe some... questionable choices in how I tried to solve the failing tests. Basically
modified how the Row tests are working so they will attempt to use the underlying layout's field
length if it gets updated, rather than using the value specified initially in the Layout. This will
allow `empty_cols_ok` to 'flex' the layout so we won't trip on the rules related to matching layouts