Set read_only false to prevent error on Windows
The read_only argument causes openpyxl to leave files open on Windows, which
results in errors during unit testing. A fix implemented in `openpyxl=3.0.6`
was supposed to have corrected this, but tests are still failing. Removing
the read_only for now, which will slow down validation of Excel files, but
should be stable on all platforms.