openpyxl bugfix and Windows tests (#60)
* Update openpyxl to re-enable windows tests

A bug in openpyxl caused unit tests to fail on Windows. The most recent patch
to the package should correct this, and allow tests on Windows to pass.

* Set read_only false to prevent error on Windows

The read_only argument causes openpyxl to leave files open on Windows, which
results in errors during unit testing. A fix implemented in `openpyxl=3.0.6`
was supposed to have corrected this, but tests are still failing. Removing
the read_only for now, which will slow down validation of Excel files, but
should be stable on all platforms.

* Set testing tempdir to session scope

Trying to figure out some Windows-only errors that are occuring during testing.

* Reverting scope change

That didn't fix it.

* Force file object to defer newline to CSV writer