Remove broken test
This was checking for keyword driven behavior which was removed some time ago,
but didn't fail the test because the expected behavior was to raise a
`TypeError` (which is what happens when you pass an invalid keyword argument
to a function.

This was exposed when `**kwargs` was added to a base class, and the `TypeError`
was no longer raised, exposing the worthless test.