Attempt to update tests to check a negative and positive case
I found that `_BaseSubject._flatten_exceptions()` was expecting an exception class to actually contain errors,
where one of the tests in this commit are explicitly trying to find cases where we do not contain that error. I
suspect there is a way better way to do this... but this seems to work!