Add exception handling to usage of tempfile.TemporaryDirectory()
I couldn't replicate this locally at all (something specific to the runner?).

I eventually found this [pytest issue](https://github.com/pytest-dev/pytest/issues/7491) that discusses
similar(ish) behaviors. This lead me to attempt to catch the PermissionError exception and essentially ignore it.

I'm not sure if this is a good approach, but it appears to resolve the issue.